fix(tools): hoist store
